Add a button to submit a new book through a modal
Used dialog tag in html to create modal and then added the event
listener to the button to open the modal and submit the form fields
to a new table row. Used a mix of grabbing the values from the input
fields while stopping the form from submitting as I did not know an
easier way.

Added some basic styling, but will likely revisit at the end of the
project.