Add a button will delete logic for one book
Added an event listener that would filter out the array by ID and then
just updated the original array to match the filtered one as I could not
think of an easier way.

Note: there is some code commented out as I forgot to commit so just
ignore until the next commit.