Refactor constructor functions into classes
Decided to make the main class Book so that I could extend that Book
creation into the Library class. I'm still a little fuzzy on what I am
doing, but the functionality seems to work so I will move forward and
continue to watch the module Youtube series when I am feeling a little
confused.